# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API - Swagger Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/admin/create": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create an Admin account",
                "operationId": "admin-create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "address",
                                    "phone_number"
                                ],
                                "properties": {
                                    "first_name": {
                                        "description": "User firstname",
                                        "type": "text"
                                    },
                                    "last_name": {
                                        "description": "User lastname",
                                        "type": "text"
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "text"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "User password",
                                        "type": "password"
                                    },
                                    "address": {
                                        "description": "User main address",
                                        "type": "text"
                                    },
                                    "phone_number": {
                                        "description": "User main phone number",
                                        "type": "text"
                                    },
                                    "is_marketing": {
                                        "description": "User marketing preferences",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "422": {
                        "description": "Validation Failed"
                    }
                }
            }
        },
        "/api/v1/admin/user-listing": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List all users",
                "operationId": "admin-user-listing",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "created_at",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "marketing",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "0",
                                "1"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "422": {
                        "description": "Validation Failed"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/v1/admin/user-edit/{uuid}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Edit a User account",
                "operationId": "admin-user-edit",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "address",
                                    "phone_number"
                                ],
                                "properties": {
                                    "first_name": {
                                        "description": "User firstname",
                                        "type": "text"
                                    },
                                    "last_name": {
                                        "description": "User lastname",
                                        "type": "text"
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "text"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "User password",
                                        "type": "password"
                                    },
                                    "address": {
                                        "description": "User main address",
                                        "type": "text"
                                    },
                                    "phone_number": {
                                        "description": "User main phone number",
                                        "type": "text"
                                    },
                                    "is_marketing": {
                                        "description": "User marketing preferences",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "422": {
                        "description": "Validation Failed"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/v1/admin/user-delete/{uuid}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete a User account",
                "operationId": "admin-user-delete",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "422": {
                        "description": "Validation Failed"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "jwt": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}